requirements.txt
amqp==2.6.1
asgiref==3.5.0
billiard==3.6.4.0
black==22.1.0
celery==4.4.7
click==8.0.4
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
filelock==3.6.0
kombu==4.6.11
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pyparsing==3.0.7
pytz==2021.3
redis==4.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
vine==1.3.0
virtualenv==20.13.2
wrapt==1.13.3
